+++ src/main/java/com/tinhocanhtrang/repositoty/ProducerRepository.java
@@ -1,0 +1,10 @@
+package com.tinhocanhtrang.repositoty;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.tinhocanhtrang.entity.Producter;
+
+
+public interface ProducerRepository extends JpaRepository<Producter, String> {
+
+}
